+++ src/components/Badge/index.js
@@ -13,22 +13,43 @@ const badgeBgs = {
   LIGHT: 'light'
 }
 
-const Badge = ({ ...props }) => {
+const Badge = ({ children, badge, className, ...props }) => {
+  if (badge) {
+    return (
+      <div className="position-relative" >
+        <BBadge
+          className={`position-absolute top-0 start-100 translate-middle badge rounded-pill ${className}`}
+          {...props}
+        >
+          {badge}
+          <span className="visually-hidden" />
+        </BBadge>
+        {children}
+      </div >
+    )
+  }
+
   return (
-    <BBadge {...props} />
+    <BBadge className={className} {...props}>
+      {children}
+    </BBadge>
   );
 };
 
 Badge.defaultProps = {
+  className: '',
   bg: badgeBgs.SECONDARY,
   pill: false,
-  text: undefined
+  text: undefined,
+  badge: undefined,
 };
 
 Badge.propTypes = {
+  className: PropTypes.string,
   bg: PropTypes.oneOf(Object.values(badgeBgs)),
   pill: PropTypes.bool,
-  text: PropTypes.oneOf(Object.values(badgeBgs))
+  text: PropTypes.oneOf(Object.values(badgeBgs)),
+  badge: PropTypes.any
 };
 
 export default Badge;