+++ src/components/index.js
@@ -28,6 +28,7 @@ export * from './Select';
 export * from './Toasts';
 export * from './Placeholders';
 export * from './Loading';
+export * from './Typography';
 // export * from './Form';
 // export * from './InputGroup';
 // export * from './Image';