+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## Changelog
 
-##### 1.4.3 - 8/8/2022
+##### 1.5.0-beta.1 - 9/8/2022
+
+- New-feature: Update Badge component (new config: badge)
+- New-feature: Add Typography component
+- Improvement: Modify Table component (Handle table empty)
+
+##### 1.4.3 - 9/8/2022
 
 - New-feature: Add loading for Button component
 - New-feature: Add scroll and pagination for Table component (config maxHeight, pagination)